2d_opengl_pro

+++ 2D_OpenGL/GLFont.h
@@ -10,10 +10,11 @@
 class CGLFont  
 {
 public:
+	// 文字栅格化的实现
 	void Printftext(int x, int y, LPCTSTR lpszText,HFONT hFont);
 	//根据字体输出文字
 	void settext(float x,float y,const char* str,int ifont,float r,float g,float b);
-	void Print2D(float x,float y,const char* str,int ifont,float r,float g,float b);
+	void Print2D(int x,int y,const char* str,int ifont,float r,float g,float b);
 	//
 	void settextTest(int x,int y,const char* str,int ifont,float r,float g,float b);
 
@@ -21,6 +22,7 @@ public:
 	virtual ~CGLFont();
 
 public:
+	// 创建字体
 	HFONT	hFontAr[16];
 };
